Rename mirror.hpp to create_mirror.hpp (#375)
