+++ include/ddc/ddc.hpp
@@ -29,8 +29,8 @@
 #include "ddc/uniform_point_sampling.hpp"
 
 // Algorithms
+#include "ddc/create_mirror.hpp"
 #include "ddc/for_each.hpp"
-#include "ddc/mirror.hpp"
 #include "ddc/parallel_deepcopy.hpp"
 #include "ddc/parallel_fill.hpp"
 #include "ddc/parallel_for_each.hpp"
